Implementacion rol bcg, proyecto servicio

+++ backend/src/main/java/com/example/backend/Controller/ProyectoController.java
@@ -1,7 +1,10 @@
 package com.example.backend.Controller;
 
+import java.security.Principal;
 import java.util.List;
 
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,7 +14,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import com.example.backend.DTO.ProyectoDTO;
-import com.example.backend.Exceptions.ValidationException;
+import com.example.backend.Models.BCG;
+import com.example.backend.Models.Role;
+import com.example.backend.Services.Implementacion.BCGServicio;
 import com.example.backend.Services.Implementacion.ProyectoServicio;
 import lombok.RequiredArgsConstructor;
 
@@ -20,35 +25,69 @@ import lombok.RequiredArgsConstructor;
 public class ProyectoController {
 
     private ProyectoServicio proyectoServicio;
+    private BCGServicio bcgServicio;
 
     @PostMapping("/proyecto/crear")
-    public ResponseEntity<String> crearProyecto(@RequestBody ProyectoDTO proyectoDTO) {
+    public ResponseEntity<String> crearProyecto(@RequestBody ProyectoDTO proyectoDTO, Principal principal) {
+
+        String username = principal.getName();
+
+        BCG bcg = bcgServicio.buscarBCG(username);
+
+        if (!bcg.getRole().equals(Role.BCG)) {
+
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("No tienes permisos para crear proyectos.");
+
+        }
+
         try {
 
             String mensajeSalida = proyectoServicio.crearProyecto(proyectoDTO);
+
             return ResponseEntity.ok(mensajeSalida);
-        } catch (RuntimeException e) {
-            return ResponseEntity.badRequest().body("Error al modificar el usuario " + e.getMessage());
+
+        } catch (DataIntegrityViolationException e) {
+
+            return ResponseEntity.badRequest().body("Error al crear el proyecto " + e.getMessage());
+
         }
     }
 
     @GetMapping("/proyecto/buscarProyectoPorNombre")
     public ResponseEntity<?> buscarProyecto(@RequestParam String nombreProyecto) {
+
         try {
-            ProyectoDTO proyectoBuscado = proyectoServicio.buscarProyecto(nombreProyecto);
+
+            ProyectoDTO proyectoBuscado = proyectoServicio.buscarProyectoDTO(nombreProyecto);
+
             return ResponseEntity.ok(proyectoBuscado);
+
         } catch (RuntimeException e) {
+
             return ResponseEntity.badRequest().body(e.getMessage());
+
         }
     }
 
     @DeleteMapping("/proyecto/eliminarProyecto")
-    public ResponseEntity<String> eliminarUsuarioGeneral(@RequestParam String nombreProyecto) {
+    public ResponseEntity<String> eliminarProyecto(@RequestParam String nombreProyecto, Principal principal) {
+        String username = principal.getName();
+
+        BCG bcg = bcgServicio.buscarBCG(username);
+
+        if (!bcg.getRole().equals(Role.BCG)) {
+
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("No tienes permisos para eliminar proyectos.");
+
+        }
+
         try {
+
             proyectoServicio.eliminarProyecto(nombreProyecto);
 
             return ResponseEntity.ok("El proyecto ha sido eliminado correctamente.");
-        } catch (IllegalArgumentException e) {
+
+        } catch (RuntimeException e) {
 
             return ResponseEntity.badRequest().body(e.getMessage());
 
@@ -56,22 +95,43 @@ public class ProyectoController {
     }
 
     @PutMapping("/proyecto/modificarProyecto")
-    public ResponseEntity<String> modificarProyecto(@RequestBody ProyectoDTO proyectoDTO) {
+    public ResponseEntity<String> modificarProyecto(@RequestBody ProyectoDTO proyectoDTO, Principal principal) {
+
         try {
+            String username = principal.getName();
+
+            BCG bcg = bcgServicio.buscarBCG(username);
+
+            if (!bcg.getRole().equals(Role.BCG)) {
+
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body("No tienes permisos para modificar proyectos.");
+
+            }
+
             String salidaModificar = proyectoServicio.modificarProyecto(proyectoDTO);
+
             return ResponseEntity.ok(salidaModificar);
-        } catch (ValidationException e) {
-            return ResponseEntity.badRequest().body("Error al modificar el usuario " + e.getMessage());
+
+        } catch (RuntimeException e) {
+
+            return ResponseEntity.badRequest().body("Error al modificar el proyecto " + e.getMessage());
+
         }
+
     }
 
     @GetMapping("/proyecto/obtenerTodosLosProyectos")
     public ResponseEntity<?> obtenerTodosLosProyectos() {
         try {
+
             List<ProyectoDTO> listaDeProyectos = proyectoServicio.obtenerTodosLosProyectos();
+
             return ResponseEntity.ok(listaDeProyectos);
+
         } catch (RuntimeException e) {
+
             return ResponseEntity.badRequest().body("Error al obtener la lista de proyectos " + e.getMessage());
+
         }
     }
 }
